# api/docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "summary": "Register a new account",
        "description": "Creates a user + organisation and returns an access token. The refresh token is set as the `crossbill_rt` httpOnly cookie.",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Account created",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "success": true,
                    "data": {
                      "accessToken": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Request validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "summary": "Log in with email & password",
        "description": "Returns an access token and sets the `crossbill_rt` httpOnly refresh cookie.",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "success": true,
                    "data": {
                      "accessToken": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Request validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
...
    "/auth/refresh": {
      "post": {
        "operationId": "AuthController_refresh",
        "summary": "Rotate tokens",
        "description": "Reads the `crossbill_rt` cookie, rotates the refresh token and returns a fresh access token.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "New access token issued",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "success": true,
                    "data": {
                      "accessToken": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Auth"
        ],
        "security": [
          {
            "crossbill_rt": []
          }
        ]
      }
    },
    "/auth/logout": {
      "post": {
        "operationId": "AuthController_logout",
        "summary": "Log out",
        "description": "Revokes refresh tokens for the user and clears the refresh cookie.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Logged out",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "success": true,
                    "data": {
                      "loggedOut": true
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated but role not permitted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Auth"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Crossbill API",
    "description": "Crossbill backend API — GST / FEMA compliant invoicing & remittance for Indian exporters.\n\n**Conventions**\n- Base path: `/api/v1`\n- Success envelope: `{ \"success\": true, \"data\": ... }` (list endpoints add a `meta` object).\n- Error envelope: `{ \"success\": false, \"error\": { \"code\", \"message\", \"correlationId\" }, \"timestamp\", \"path\" }`.\n- Auth: send `Authorization: Bearer <accessToken>` (obtained from `/auth/login` or `/auth/register`).\n- The refresh token is delivered as the `crossbill_rt` httpOnly cookie and is read automatically by `/auth/refresh`.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:4000/api/v1",
      "description": "Base URL (includes /api/v1 prefix)"
    }
  ],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Paste the accessToken returned by /auth/login or /auth/register."
      },
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "crossbill_rt",
        "description": "httpOnly refresh-token cookie set by the auth endpoints; used by /auth/refresh."
      }
    },
    "schemas": {
      "RegisterDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "Email address to register the account under (must be unique)",
            "format": "email",
            "example": "founder@acme.in"
          },
          "password": {
            "type": "string",
            "description": "Password, minimum 8 characters",
            "example": "S3curePass!",
            "minLength": 8
          },
          "legalName": {
            "type": "string",
            "description": "Registered legal/business name used on invoices",
            "example": "Acme Software Exports Pvt Ltd"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "ApiErrorBody": {
        "type": "object",
        "properties": {
          "code": {
            "type": "number",
            "example": 400,
            "description": "HTTP status code"
          },
          "message": {
            "type": "string",
            "example": "Validation failed",
            "description": "Human-readable error message"
          },
          "correlationId": {
            "type": "string",
            "example": "7f3c1e9a-2b4d-4c8e-9a1b-1f2e3d4c5b6a",
            "description": "Request correlation id"
          }
        },
        "required": [
          "code",
          "message",
          "correlationId"
        ]
      },
      "ApiErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "$ref": "#/components/schemas/ApiErrorBody"
          },
          "timestamp": {
            "type": "string",
            "example": "2026-06-12T10:15:30.000Z"
          },
          "path": {
            "type": "string",
            "example": "/api/v1/clients"
          }
        },
        "required": [
          "success",
          "error",
          "timestamp",
          "path"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "Account email address",
            "format": "email",
            "example": "founder@acme.in"
          },
          "password": {
            "type": "string",
            "description": "Account password",
            "example": "S3curePass!",
            "minLength": 8
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
...
    }
  }
}